PortfolioConsumer/requirements.txt
amqp==5.0.6
aniso8601==9.0.1
billiard==3.6.4.0
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
Flask-RESTful==0.3.9
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
numpy
pandas
prompt-toolkit==3.0.18
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
kafka-python==2.0.2
redis==3.5.3
flask_sqlalchemy
pytest
python-json-logger
python3-logstash